+++ test/Feature/StructureSpec.hs
@@ -231,6 +231,23 @@ spec = do
               ]
             |]
 
+    describe "VIEW that has a source FK based on a UNIQUE key" $
+
+      it "includes fk description" $ do
+        r <- simpleBody <$> get "/"
+
+        let referralLink = r ^? key "definitions" . key "referrals" . key "properties" . key "link"
+
+        liftIO $
+          referralLink `shouldBe` Just
+            [aesonQQ|
+              {
+                "format": "integer",
+                "type": "integer",
+                "description": "Note:\nThis is a Foreign Key to `pages.link`.<fk table='pages' column='link'/>"
+              }
+            |]
+
     describe "PostgreSQL to Swagger Type Mapping" $ do
 
       it "character varying to string" $ do